[Chore #161947401] Arranging files
