+++ app/users.py
@@ -5,6 +5,19 @@ from functools import wraps
 from __init__ import *
 import jwt
 
+users =  {}
+
+def mustlogin(d):
+	@wraps(d)
+	def decorated(*args, **kwargs):
+		if request.headers.get('	')=='':
+			return make_response(("You need to first login "), 201)
+		try:
+			jwt.decode(request.headers.get('x-access-token'), "djrefuge")
+		except:
+			return jsonify({"message": 'please sigin '})
+		return d(*args, **kwargs)
+	return decorated 
 
 class Register(Resource):
 	def post(self):
