--- login.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<title>SendIT</title>
-	<link rel="stylesheet" type="text/css" href="css/style.css">
-</head>
-<body>
-<div class="top">
-<img src="img/delivery.png">
-  <a href="login.html">Login</a>
-  <a href="register.html">Register</a>
-   <a href="index.html">Home</a>  
-
-  </div> 
-
-
-<h2 style="text-align: center; font-size: 25px;">Login</h2>
-<input type='checkbox' id='form-switch'>
-
-<form id='login-form' action="dashboard.html" method='post'>
-  <input type="text" placeholder="Username" required>
-  <input type="password" placeholder="Password" required>
-  <button type='submit'>Login</button>
-  <label for='form-switch' style="color: #0099CC; padding-left: 50px;"><span>Forget Password</span></label><br/><br/>
-    <a href="register.html">Don't have account? Sigup</a>
-</form>
-<form id='register-form' action="" method='post'>
-  <input type="email" placeholder="Email" required>
-  <button type='submit'>Send</button>
- 
-</form>
-</body>
-</html>
